visualization/requirements.txt
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.21.0
cartopy>=0.21.0
pyyaml>=6.0
